requirements.txt
# Core Frameworks
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# Image Processing
opencv-python
Pillow

# 3D Processing & Visualization
open3d>=0.17.0

# Model Libraries
ultralytics                 # For YOLO-World
transformers                # For Depth Anything
accelerate                  # Required for efficient inference with Transformers

# Segment Anything (Official Repo)
git+https://github.com/facebookresearch/segment-anything.git

# Optional for specialized math
scipy